+++ clients/java/signalr/src/main/java/com/microsoft/aspnet/signalr/HttpConnectionOptions.java
@@ -9,6 +9,8 @@ public class HttpConnectionOptions {
     private LogLevel loglevel;
     private boolean skipNegotiate;
 
+    public HttpConnectionOptions() {}
+
     public HttpConnectionOptions(String url, Transport transport, LogLevel logLevel, boolean skipNegotiate) {
         this.url = url;
         this.transport = transport;
